move tracklist request from artistList to artist Card to proper render
