+++ src/actions/index.js
@@ -5,6 +5,8 @@ const SET_ARTIST_PARAM = 'SET_ARTIST_PARAM'
 const SET_TRACK_PARAM = 'SET_TRACK_PARAM'
 const SET_MODAL = 'SET_MODAL';
 
+const SET_TRACKLIST = 'SET_TRACKLIST';
+
 const getArtistList = artistList => ({
   type: GET_ARTIST_LIST,
   artistList,
@@ -35,9 +37,14 @@ const setModal = modal => ({
   modal,
 })
 
+const settrackList = trackList => ({
+  type: SET_TRACKLIST,
+  trackList,
+})
+
 export {
   GET_ARTIST_LIST, SET_CURRENT_CARD, SET_PAGINATION, SET_ARTIST_PARAM,
-  SET_TRACK_PARAM, SET_MODAL,
+  SET_TRACK_PARAM, SET_MODAL, SET_TRACKLIST,
   getArtistList, setCurrentCard, setPagination, setArtistParam,
-  setTrackParam, setModal,
+  setTrackParam, setModal, settrackList,
 };
