+++ src/components/ArtistCard.js
@@ -1,15 +1,28 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { useDispatch } from 'react-redux';
-import { setCurrentCard, setModal } from '../actions/index';
+import { setCurrentCard, setModal, settrackList } from '../actions/index';
+import { trackListApi } from '../services/discogsApi';
+import { toast } from 'react-toastify';
 
 const ArtistCard = props => {
   const { id, thumb, title, year, card } = props;
   const dispatch = useDispatch();
+  let trackList = [];
 
   const handleClick = () => {
     dispatch(setCurrentCard(card));
-    dispatch(setModal(true));
+
+    trackListApi(card.master_url)
+      .then(response => {
+        if (response.error) {
+          toast.error(response.error);
+        } else {
+          response.tracklist.map(track => trackList.push(track.title));
+          dispatch(settrackList(trackList));
+          dispatch(setModal(true));
+        }
+      })
   }
 
   return (
