+++ src/containers/ArtistList.js
@@ -6,7 +6,6 @@ import { discogsApi, updateApi } from '../services/discogsApi';
 import ArtistCard from '../components/ArtistCard';
 import ReactPaginate from 'react-paginate';
 import Details from '../components/Details';
-import { trackListApi } from '../services/discogsApi';
 
 toast.configure();
 const ArtistList = () => {
@@ -15,9 +14,6 @@ const ArtistList = () => {
   const artist = useSelector(state => state.setParams.artist)
   const track = useSelector(state => state.setParams.track)
   const dataPagination = useSelector(state => state.dataPagination)
-
-  let trackList = [];
-  const currentCard = useSelector(state => state.artistList.currentCard);
 
   useEffect(() => {
     if (artistList && artistList.length === 0 ) {
@@ -43,17 +39,6 @@ const ArtistList = () => {
           dispatch(getArtistList(response.results))
         }
       })
-  }
-
-  if (currentCard.length !== 0) {
-    trackListApi(currentCard.master_url)
-    .then(response => {
-      if (response.error) {
-        toast.error(response.error);
-      } else {
-        response.tracklist.map(track => trackList.push(track.title));
-      }
-    })
   }
 
   const handlePageClick = (e) => {
@@ -88,7 +73,7 @@ const ArtistList = () => {
             />
         ))) : <div>Loading</div>}
       </div>
-      <Details trackList={trackList} />
+      <Details />
     </div>
   )
 }
