+++ src/reducers/artistList.js
@@ -1,8 +1,9 @@
-import { GET_ARTIST_LIST, SET_CURRENT_CARD } from '../actions';
+import { GET_ARTIST_LIST, SET_CURRENT_CARD, SET_TRACKLIST } from '../actions';
 
 const initialState = {
   artistList: [],
   currentCard: [],
+  trackList: [],
 };
 
 const artistListReducer = (state = initialState, action) => {
@@ -16,6 +17,11 @@ const artistListReducer = (state = initialState, action) => {
       return {
         ...state,
         currentCard: action.currentCard,
+      };
+    case SET_TRACKLIST:
+      return {
+        ...state,
+        trackList: action.trackList,
       };
     default:
       return state;
